code corrected
