+++ model/users/usersController.js
@@ -1,10 +1,6 @@
 const jwt = require('jsonwebtoken')
 const Users = require('./usersModel')
-const fs = require('fs').promises
-const path = require('path')
-const Jimp = require('jimp')
 const { HttpCode } = require('../../helpers/constants')
-const createFolderIsExist = require('../../helpers/create-dir')
 require('dotenv').config()
 const SECRET_KEY = process.env.JWT_SECRET
 
@@ -112,28 +108,7 @@ const updateSub = async (req, res, next) => {
 const avatars = async (req, res, next) => {
   try {
     const id = req.user.id
-    const AVATARS_OF_USERS = process.env.AVATARS_OF_USERS
-    const pathFile = req.file.path
-    const newNameAvatar = `${Date.now()}-${req.file.originalname}`
-    const img = await Jimp.read(pathFile)
-    img
-      .autocrop()
-      .cover(
-        250,
-        250,
-        Jimp.HORIZONTAL_ALIGN_CENTER | Jimp.VERTICAL_ALIGN_MIDDLE,
-      )
-      .writeAsync(pathFile)
-    await createFolderIsExist(path.join(AVATARS_OF_USERS, id))
-    await fs.rename(pathFile, path.join(AVATARS_OF_USERS, id, newNameAvatar))
-    const avatarUrl = path.normalize(path.join(id, newNameAvatar))
-    try {
-      await fs.unlink(
-        path.join(process.cwd(), AVATARS_OF_USERS, req.user.avatarUrl),
-      )
-    } catch (err) {
-      console.log(err.message)
-    }
+    const avatarUrl = await Users.saveAvatarToStatick(req)
     await Users.updateAvatar(id, avatarUrl)
     return res.status(HttpCode.OK).json(avatarUrl)
   } catch (err) {
