+++ model/users/usersModel.js
@@ -1,4 +1,8 @@
+const fs = require('fs').promises
+const path = require('path')
+const Jimp = require('jimp')
 const User = require('./usersSchema')
+const createFolderIsExist = require('../../helpers/create-dir')
 
 const findUserByEmail = async email => {
   return await User.findOne({ email })
@@ -29,6 +33,29 @@ const updateAvatar = async (id, avatarUrl) => {
   return await User.updateOne({ _id: id }, { avatarUrl: avatarUrl })
 }
 
+const saveAvatarToStatick = async req => {
+  const id = req.user.id
+  const AVATARS_OF_USERS = process.env.AVATARS_OF_USERS
+  const pathFile = req.file.path
+  const newNameAvatar = `${Date.now()}-${req.file.originalname}`
+  const img = await Jimp.read(pathFile)
+  img
+    .autocrop()
+    .cover(250, 250, Jimp.HORIZONTAL_ALIGN_CENTER | Jimp.VERTICAL_ALIGN_MIDDLE)
+    .writeAsync(pathFile)
+  await createFolderIsExist(path.join(AVATARS_OF_USERS, id))
+  await fs.rename(pathFile, path.join(AVATARS_OF_USERS, id, newNameAvatar))
+  const avatarUrl = path.normalize(path.join(id, newNameAvatar))
+  try {
+    await fs.unlink(
+      path.join(process.cwd(), AVATARS_OF_USERS, req.user.avatarUrl),
+    )
+  } catch (err) {
+    console.log(err.message)
+  }
+  return avatarUrl
+}
+
 module.exports = {
   findUserByEmail,
   findUserById,
@@ -37,4 +64,5 @@ module.exports = {
   updateUserSub,
   findUserByToken,
   updateAvatar,
+  saveAvatarToStatick,
 }
